damn now it works
