--- frontend/app/.env
@@ -1,4 +1,0 @@
-NEXT_PUBLIC_ALCHEMY_ID=Lb71sm_wwKUpO6DitJ7WA5pPX86DgPme
-NEXT_PUBLIC_INFURA_ID=
-NEXT_PUBLIC_WALLETCONNECT_PROJECT_ID="3f296b8d0eb4eb510d067de47f8ab560"
-SESSION_SECRET=
