--- frontend/app/components/ConnectWalletButton.tsx
@@ -1,6 +1,0 @@
-"use client";
-import { ConnectKitButton } from "connectkit";
-
-export default function ConnectWalletButton() {
-  return <ConnectKitButton />;
-}
