--- frontend/app/layout.tsx
@@ -1,28 +1,0 @@
-// import type { ReactNode } from 'react';
-// import { Inter } from "next/font/google";
-// import ConnectWalletButton from './components/ConnectWalletButton';
-// import MyComponent from './components/MyComponent';
-
-// const inter = Inter({ subsets: ["latin"] });
-
-// type RootLayoutProps = {
-//   children: ReactNode;
-// };
-
-// const RootLayout = ({ children }: RootLayoutProps) => {
-//   return (
-//     <div className={inter.className}>
-//       <header>
-//         {/* Header content */}
-//       </header>
-//       <ConnectWalletButton />
-//       <MyComponent />
-//       {children}
-//       <footer>
-//         {/* Footer content */}
-//       </footer>
-//     </div>
-//   );
-// };
-
-// export default RootLayout;
