+++ frontend/next.config.js
@@ -1,6 +1,6 @@
 module.exports = {
-    webpack: (config) => {
-      config.resolve.fallback = { fs: false, net: false, tls: false };
-      return config;
-    },
-  };
+  webpack: (config) => {
+    config.resolve.fallback = { fs: false, net: false, tls: false };
+    return config;
+  },
+};
