+++ frontend/pages/components/ ConnectedWallet.tsx
@@ -1,18 +1,10 @@
 "use client";
-
 import { useAccount } from "wagmi";
 
+// Make sure that this component is wrapped with ConnectKitProvider
 const MyComponent = () => {
   const { address, isConnecting, isDisconnected } = useAccount();
-
-  if (isConnecting) {
-    return <div>Connecting...</div>;
-  }
-  if (isDisconnected) {
-    return <div>Disconnected</div>;
-  }
-
+  if (isConnecting) return <div>Connecting...</div>;
+  if (isDisconnected) return <div>Disconnected</div>;
   return <div>Connected Wallet: {address}</div>;
 };
-
-export default MyComponent;
