+++ frontend/pages/index.tsx
@@ -1,4 +1,3 @@
-"use client";
 import type { NextPage } from 'next';
 import { ConnectKitButton } from 'connectkit';
